Moved functions to separate file & combined interpolation
1) Functions from both the pre-processing.py & the vemamodeling.py files were moved into a separate file (modelling_functions.py)
2) 3 different interpolation functions (add_history_and_future, stock_tail & lifetime_tail) were combined into one generic interpolation function (interpolate)
